Make Rational never promote; update intersection/incircle defaults and tooling
Rational manages its own overflow by reducing to lowest terms, so promoting
the whole fraction type was redundant. Remove the Rational _promote
specializations so promoted_number_t<Rational<T>> == Rational<T>, and update
the corresponding test.

Follow-on adjustments now that Rational stays at its chosen width:
- intersections.hpp: default the Bentley-Ottmann exactness type to
  Rational<BigInt> instead of Rational<int128>, so the algorithms stay exact
  without relying on promotion.
- orientation.hpp: double-promote the inCircle coordinate (its degree-4
  determinant needs more headroom than the orientation test), e.g.
  int64 -> int128 -> BigInt.

Tooling and docs:
- doctable.cpp: skip the promotion-enabled benchmark for types whose promoted
  type equals the original (BigInt, long double, every Rational), avoiding
  redundant and slow runs.
- make_type_tables.py: new script that compiles, runs and formats the
  doc/types.md benchmark tables; leaves promotion cells empty for
  never-promoted types and drops the always-empty promotion "/ 60" column.
- boosttable.cpp, doc/types.md: matching benchmark/documentation updates.

+++ tests/unit/rational.cpp
@@ -84,7 +84,8 @@ TEST_CASE("Rational numeric limits and promotion preserve rational types") {
     using SmallRational = pgl::Rational<int16_t>;
     using PromotedSmallRational = pgl::detail::promoted_number_t<SmallRational>;
 
-    static_assert(std::is_same_v<PromotedSmallRational, pgl::Rational<int32_t>>);
+    // Rational is never promoted: the storage type is preserved.
+    static_assert(std::is_same_v<PromotedSmallRational, pgl::Rational<int16_t>>);
 
     CHECK(std::numeric_limits<pgl::Rational<int>>::min() == pgl::Rational<int>(std::numeric_limits<int>::min()));
     CHECK(std::numeric_limits<pgl::Rational<int>>::max() == pgl::Rational<int>(std::numeric_limits<int>::max()));
